Repo Men Inc - lesson 3, step 2

+++ team_repo_men_inc/src/main/java/lv/javaguru/java2/repo_men_inc/services/RemoveDebtorService.java
@@ -1,17 +1,29 @@
 package lv.javaguru.java2.repo_men_inc.services;
 
 import lv.javaguru.java2.repo_men_inc.core.requests.RemoveDebtorRequest;
+import lv.javaguru.java2.repo_men_inc.core.responses.CoreError;
 import lv.javaguru.java2.repo_men_inc.core.responses.RemoveDebtorResponse;
+import lv.javaguru.java2.repo_men_inc.core.validators.RemoveDebtorValidator;
 import lv.javaguru.java2.repo_men_inc.database.Database;
 
-public class RemoveDebtorService {
-    Database database;
+import java.util.List;
 
-    public RemoveDebtorService(Database database) {
+public class RemoveDebtorService {
+    private Database database;
+    private RemoveDebtorValidator removeDebtorValidator;
+
+    public RemoveDebtorService(Database database, RemoveDebtorValidator removeDebtorValidator) {
         this.database = database;
+        this.removeDebtorValidator = removeDebtorValidator;
     }
 
     public RemoveDebtorResponse execute(RemoveDebtorRequest removeDebtorRequest) {
+
+        List<CoreError> errors = removeDebtorValidator.validate(removeDebtorRequest);
+        if (!errors.isEmpty()) {
+            return new RemoveDebtorResponse(errors);
+        }
+
         boolean isDebtorRemoved = database.deleteById(removeDebtorRequest.getDebtorsId());
         return new RemoveDebtorResponse(isDebtorRemoved);
     }
